backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
greenlet==3.1.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.19
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==11.4.1
sendgrid==6.11.0
websockets==14.1
python-dateutil==2.9.0
pillow==11.1.0
aiofiles==24.1.0
